requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
pandas
python-jose[cryptography]
passlib[bcrypt]
httpx
python-dotenv
